fix default page parameter handling for API
If no page parameter is specified (e.g. if it is nil or an empty string)
we should default to a value instead of returning a 500.
This picks the 'all_pages' behaviour as the default.

The original issue can be reproduced by navigating to any app, clicking
"callers" and then clicking on "API response".