fix timezone issues for DST
Using Time.now.zone returns "CEST" which is not valid
for in_time_zone as that expects a "Continent/City" string
used by the IANA.

This uses localtime, which drops the timezone name, but keeps the correct
UTC offset.

Another solution would be to explicitly set the timezone in Rails
via config.time_zone
https://api.rubyonrails.org/classes/ActiveSupport/TimeZone.html#method-c-new

But we would need to either set it ourselves or read it from somewhere, which
does not seem necessary in this case as the UTC offset should be enough.